Move files and folder (ContractUsPage, Quote
