+++ src/components/Layouts/MainLayout.tsx
@@ -1,6 +1,6 @@
 import Header from "@components/Layouts/Header";
 import Footer from "@components/Layouts/Footer";
-import Quote from "@components/Layouts/Quote";
+import Quote from "@components/Base/Quote";
 import ScrollButton from "@components/Base/ScrollButton";
 
 export default function MainLayout({children}) {
