--- src/components/Pages/ExhibitionPage/Projects/ProjectPage/Carousel/Carousel.tsx
@@ -1,1 +1,0 @@
-export default function Carousel() {}
