+++ src/components/Pages/ExhibitionPage/Projects/ProjectPage/Carousel/index.tsx
@@ -1,0 +1,3 @@
+export default function Carousel() {
+  return <div></div>;
+}
